use live data for article list
